btbucket - bug fix for 64bit machine

+++ src/btbucket.c
@@ -19,7 +19,7 @@
 #include <limits.h>
 
 void btbucket_new(void *btbucket, unsigned int bucketsize, 
-  unsigned int sib_fileno, unsigned int sib_offset, int leaf, void *prefix, 
+  unsigned int sib_fileno, unsigned long int sib_offset, int leaf, void *prefix, 
   unsigned int *prefix_size) {
     struct btbucket *b = btbucket;
